Files renamed and organized appropriately
