+++ GCCHRConnect/GCCHRMachinery/GCCHRMachinery.csproj
@@ -68,7 +68,7 @@
     <Compile Include="Entities\IMongoEntity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utilities\Formatter.cs" />
-    <Compile Include="Utilities\StringCipher.cs" />
+    <Compile Include="Security\Cryptography.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
