+++ GCCHRConnect/GCCHRMachineryTest/GCCHRMachineryTest.csproj
@@ -59,13 +59,16 @@
     <Compile Include="DataAccessLayer\MongoDb\Contacts management\ContactDBTest.cs" />
     <Compile Include="DataAccessLayer\MongoDb\Contacts management\TagDBTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Utilities\StringCipherTest.cs" />
+    <Compile Include="Security\CryptographyTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\GCCHRMachinery\GCCHRMachinery.csproj">
       <Project>{6230dd14-facb-4060-8e2b-82df38c53441}</Project>
       <Name>GCCHRMachinery</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Utilities\" />
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
